Renamed rocket.h/.cpp to rocketFSM.h/.cpp
