+++ flight/mcu/lib/teensyShared/rocketFSM.cpp
@@ -1,4 +1,4 @@
-#include "rocket.h"
+#include "rocketFSM.h"
 
 #include <Arduino.h>
 #include <ChRt.h>
