+++ flight/mcu/platformio.ini
@@ -34,7 +34,7 @@ lib_deps =
   googletest
 build_flags =
   -pthread 
-src_filter = -<mcu_hybrid/> -<mcu_ac>
+src_filter = -<mcu_hybrid/> -<mcu_ac/>
 test_ignore = target
